+++ CHANGES.rst
@@ -11,15 +11,19 @@ Version 1.14
 
 Unreleased
 
-* Use ``PRAGMA user_version`` instead of a version table. (:issue:`210`)
-* Use ``PRAGMA application_id`` to identify reader databases;
-  the id is ``0x66656564`` – ``read`` in ASCII / UTF-8. (:issue:`211`)
 * Add the :meth:`~Reader.update_feeds_iter` method,
   which yields the update status of each feed as it gets updated.
   (:issue:`204`)
 * Change the return type of :meth:`~Reader.update_feed`
   from ``None`` to ``Optional[UpdatedFeed]``.
   (:issue:`204`)
+* Add the ``session_timeout`` argument to :func:`make_reader`
+  to set a timeout for retrieving HTTP(S) feeds.
+  The default (connect timeout, read timeout) is (3.05, 60) seconds;
+  the previous behavior was to *never time out*.
+* Use ``PRAGMA user_version`` instead of a version table. (:issue:`210`)
+* Use ``PRAGMA application_id`` to identify reader databases;
+  the id is ``0x66656564`` – ``read`` in ASCII / UTF-8. (:issue:`211`)
 * Change the ``reader update`` command to show a progress bar
   and update summary (with colors), instead of plain log output.
   (:issue:`204`)